+++ src/test/java/egovframework/com/cop/cmt/service/impl/EgovArticleCommentDAOTest_selectArticleCommentListCnt.java
@@ -1,4 +1,4 @@
-package egovframework.com.cop.bbs.service.impl;
+package egovframework.com.cop.cmt.service.impl;
 
 import static org.junit.Assert.assertEquals;
 
@@ -13,7 +13,10 @@ import egovframework.com.cmm.LoginVO;
 import egovframework.com.cmm.util.EgovUserDetailsHelper;
 import egovframework.com.cop.bbs.service.Board;
 import egovframework.com.cop.bbs.service.BoardMaster;
+import egovframework.com.cop.bbs.service.impl.EgovArticleDAO;
+import egovframework.com.cop.bbs.service.impl.EgovBBSMasterDAO;
 import egovframework.com.cop.cmt.service.Comment;
+import egovframework.com.cop.cmt.service.CommentVO;
 import egovframework.com.cop.cmt.service.impl.EgovArticleCommentDAO;
 import egovframework.com.test.EgovTestV1;
 import egovframework.rte.fdl.cmmn.exception.FdlException;
@@ -23,7 +26,7 @@ import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @ContextConfiguration(classes = { EgovArticleCommentDAOTest_Configuration.class })
-public class EgovArticleCommentDAOTest_updateArticleComment extends EgovTestV1 {
+public class EgovArticleCommentDAOTest_selectArticleCommentListCnt extends EgovTestV1 {
 
 	@Resource(name = "egovBBSMstrIdGnrService")
 	private EgovIdGnrService egovBBSMstrIdGnrService;
@@ -48,7 +51,6 @@ public class EgovArticleCommentDAOTest_updateArticleComment extends EgovTestV1 {
 
 	// testData
 	String today;
-	String today2;
 	LoginVO authenticatedUser;
 
 	BoardMaster boardMaster;
@@ -58,8 +60,10 @@ public class EgovArticleCommentDAOTest_updateArticleComment extends EgovTestV1 {
 	// given
 	Comment comment;
 
+	CommentVO commentVO;
+
 	// when
-	boolean result = false;
+	int articleCommentListCnt;
 
 	@Test
 //	@Commit
@@ -123,31 +127,19 @@ public class EgovArticleCommentDAOTest_updateArticleComment extends EgovTestV1 {
 	}
 
 	void given() {
-		try {
-			Thread.sleep(1000);
-		} catch (InterruptedException e) {
-			log.error(e.getMessage());
-		}
-
-		today2 = " " + EgovDateUtil.toString(new Date(), null, null);
-		comment.setCommentCn("test 댓글" + today2);
-		comment.setLastUpdusrId(authenticatedUser.getUniqId());
-//		comment.setCommentNo("");
+		commentVO = new CommentVO();
+		commentVO.setBbsId(boardMaster.getBbsId());
+		commentVO.setNttId(board.getNttId());
 	}
 
 	void when() {
-		try {
-			egovArticleCommentDAO.updateArticleComment(comment);
-			result = true;
-		} catch (Exception e) {
-			log.error(e.getMessage());
-		}
+		articleCommentListCnt = egovArticleCommentDAO.selectArticleCommentListCnt(commentVO);
 	}
 
 	void then() {
-		log.debug("result={}", result);
+		log.debug("articleCommentListCnt={}", articleCommentListCnt);
 
-		assertEquals(result, true);
+		assertEquals(articleCommentListCnt, 1);
 	}
 
 }
